@@ -23,21 +23,12 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-  english_words: ^3.1.5
-  amazon_cognito_identity_dart:     
-    git:
-      url: git://github.com/LudySu/amazon-cognito-identity-dart.git
-      ref: master # can be a branch or a tag
   firebase_core: ^0.4.0+9
   firebase_auth: ^0.14.0+5
   provider: ^4.0.1
   flutter_spinkit: ^4.1.1+1
   cloud_firestore: ^0.13.0
-  webview_flutter: ^0.3.19+5
-  http: ^0.12.0+4
-  page_view_indicators: ^1.3.1
-  sqflite: ^1.2.0
-  path_provider: ^1.6.0
+  getflutter: ^1.0.11
 
 
 
