@@ -1,7 +1,11 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:getflutter/components/avatar/gf_avatar.dart';
+import 'package:getflutter/components/rating/gf_rating.dart';
 import 'package:omnus/Components/SearchBar.dart';
 import 'package:omnus/Firestore/SearchFunctions.dart';
+import 'package:omnus/Firestore/UserFunctions.dart';
+import 'package:omnus/MainScreens/ChefDetailsScreen.dart';
 //import 'package:omnus/Auth/AuthFunctions.dart';
 import 'package:omnus/Models/Chef.dart';
 import 'package:omnus/Models/User.dart';
@@ -13,64 +17,63 @@ class HomeScreen extends StatefulWidget {
 }
 
 class _HomeScreenState extends State<HomeScreen> {
-
   bool searching = false;
   String query = "";
   Chef hey = new Chef(id: 'sjkjkd', rate: 6, name: 'yummy', menu: {});
 
-  Future<List<Chef>> getChefs(String search) async {
-    List<Chef> chefs = List<Chef>();
-    await SearchFunctions().getAllChefs().then((result) => {
-          for (DocumentSnapshot snap in result.documents)
-            chefs.add(Chef.fromFirestore(snap))
-        });
-
-    return chefs;
-  }
-
-  void showChefPopup(BuildContext context, Chef chef){
-
+  //Unused
+  void showChefPopup(BuildContext context, Chef chef) {
     double width = MediaQuery.of(context).size.width;
     double height = MediaQuery.of(context).size.height;
 
     Dialog chefPopup = Dialog(
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(0)),
-        child: Container(
-          height: height * .4,
-          width: width*.9,
-          child: Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-              children: <Widget>[
-                Text(chef.name),
-                FlatButton(
-                  onPressed: null, 
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(0)),
+      child: Container(
+        height: height * .4,
+        width: width * .9,
+        child: Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: <Widget>[
+              Text(chef.name),
+              FlatButton(
+                  onPressed: null,
                   child: Container(
                     color: Colors.blue,
-                    height: height*.05,
+                    height: height * .05,
                     width: double.infinity,
                     alignment: Alignment.center,
-                    child: Text('Message ${chef.name}', 
+                    child: Text(
+                      'Message ${chef.name}',
                       style: TextStyle(
                         color: Colors.white,
                         fontSize: 18,
                       ),
                       textAlign: TextAlign.center,
                     ),
-                  )
-                )
-              ],
-            ),
+                  ))
+            ],
           ),
         ),
-      );
+      ),
+    );
 
-    showDialog(context: context, builder : (BuildContext context) => chefPopup);
+    showDialog(context: context, builder: (BuildContext context) => chefPopup);
   }
 
   Widget title = Text('Hello');
   Widget leading = Icon(Icons.search);
+
+  String chefType(String type) {
+    if (type == 'both') {
+      return 'Chef & Prep';
+    } else if (type == 'chef') {
+      return 'Chef';
+    } else {
+      return 'Prep';
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -82,24 +85,108 @@ class _HomeScreenState extends State<HomeScreen> {
         user = User.fromFirestore(snapshot);
       }
     }
-    
+
     if (user != null) {
       return Scaffold(
         appBar: AppBar(
-          title: Text('Hello ${user.firstName}', style: TextStyle(color: Colors.black),),
+          title: Text(
+            'Hello ${user.firstName}',
+            style: TextStyle(color: Colors.black),
+          ),
           backgroundColor: Colors.white,
           actions: <Widget>[
             IconButton(
-              icon: Icon(Icons.search, color: Colors.black,), 
-              onPressed: () => showSearch(context: context, delegate: SearchBar())
-            )
+                icon: Icon(
+                  Icons.search,
+                  color: Colors.black,
+                ),
+                onPressed: () =>
+                    showSearch(context: context, delegate: SearchBar()))
           ],
-        ),    
+        ),
+        body: FutureBuilder<QuerySnapshot>(
+            future: SearchFunctions().getAllChefs(),
+            builder: (context, snapshot) {
+              if (snapshot.hasData) {
+                List<Chef> chefs = [];
+                for (DocumentSnapshot snap in snapshot.data.documents)
+                  chefs.add(Chef.fromFirestore(snap));
+                chefs.sort((a, b) => b.numReviews.compareTo(a.numReviews));
+                return GridView.count(
+                  crossAxisCount: 1,
+                  children: List.generate(chefs.length, (index) {
+                    return Padding(
+                      padding: const EdgeInsets.all(8),
+                      child: Card(
+                        child: InkWell(
+                          onTap: () => Navigator.push(
+                            context, 
+                            MaterialPageRoute(builder: (context) => ChefDetailsScreen(chef: chefs[index]))),
+                          child: Column(
+                            children: <Widget>[
+                              Expanded(
+                                flex: 2,
+                                child: Container(
+                                  color: Colors.blueAccent,
+                                ),
+                              ),
+                              Expanded(
+                                child: Padding(
+                                  padding: const EdgeInsets.all(8.0),
+                                  child: Container(
+                                    child: Column(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.center,
+                                      crossAxisAlignment:
+                                          CrossAxisAlignment.start,
+                                      children: <Widget>[
+                                        Text(
+                                          chefs[index].name,
+                                          style: TextStyle(
+                                            fontWeight: FontWeight.bold,
+                                            fontSize: 25,
+                                          ),
+                                          textAlign: TextAlign.left,
+                                        ),
+                                        Padding(
+                                            padding: EdgeInsets.symmetric(
+                                                vertical: 1)),
+                                        Row(
+                                          children: <Widget>[
+                                            Text(
+                                              chefType(chefs[index].type),
+                                              style: TextStyle(
+                                                fontSize: 20,
+                                              ),
+                                            ),
+                                            GFRating(
+                                              allowHalfRating: true,
+                                              value: chefs[index].rating,
+                                              color: Colors.amber,
+                                              borderColor: Colors.amber,
+                                              size: 25,
+                                            )
+                                          ],
+                                        ),
+                                      ],
+                                    ),
+                                  ),
+                                ),
+                              )
+                            ],
+                          ),
+                        ),
+                      ),
+                    );
+                  }),
+                );
+              } else {
+                return Text("Loading");
+              }
+            }),
       );
     } else {
-      return Scaffold(
-        body: Center(child: Text('loading'))
-      );
+      return Center(child: Text('Loading'));
     }
   }
 }
